+++ internal/format/format.go
@@ -21,7 +21,6 @@
 package format
 
 import (
-	"github.com/uber/prototool/internal/settings"
 	"github.com/uber/prototool/internal/text"
 	"go.uber.org/zap"
 )
@@ -33,7 +32,7 @@ type Transformer interface {
 	// Failures should never happen in the CLI tool as we run the files
 	// through protoc first, but this is done because we want to verify
 	// code correctness here and protect against the bad case.
-	Transform(config settings.Config, filename string, data []byte) ([]byte, []*text.Failure, error)
+	Transform(filename string, data []byte) ([]byte, []*text.Failure, error)
 }
 
 // TransformerOption is an option for a new Transformer.
@@ -48,11 +47,11 @@ func TransformerWithLogger(logger *zap.Logger) TransformerOption {
 	}
 }
 
-// TransformerWithUpdateFileOptions returns a TransformerOption that will update the file options
+// TransformerWithRewrite returns a TransformerOption that will update the file options
 // go_package, java_package to match the package per the guidelines of the style guide.
-func TransformerWithUpdateFileOptions() TransformerOption {
+func TransformerWithRewrite() TransformerOption {
 	return func(transformer *transformer) {
-		transformer.updateFileOptions = true
+		transformer.rewrite = true
 	}
 }